Create codeql.yml
The default configuration includes `dist/`, which is generated by `ncc` from our code, our dependencies, and other sources. We do not control the output. I may explore other alternatives (I am liking `pkgroll`, and if it can guarantee me a single-file like `ncc` does, I will probably switch to it), but it is not clear to me whether this would in fact resolve the overall issue.

This is the default configuration with an explicit ignore of `dist/`.

Signed-off-by: Austin Ziegler <aziegler@kineticcommerce.com>